pyproject.toml
[tool.poetry]
name = "streamlit-workout-app"
version = "0.1.0"
description = "AI enabled Streamlit workout app"
authors = ["Jerry McCommas <jerry.mccommas@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
pandas = "^2.2.3"
watchdog = "^6.0.0"
openai = "^1.58.1"
anthropic = "^0.42.0"
python-dotenv = "^1.0.1"
streamlit-pydantic = "^0.6.0"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[server]
headless = true
